+++ tests/Integration/AuthTestCase.php
@@ -14,6 +14,7 @@ use Kreait\Firebase\Auth\SendActionLink\FailedToSendActionLink;
 use Kreait\Firebase\Auth\SignIn\FailedToSignIn;
 use Kreait\Firebase\Auth\UserRecord;
 use Kreait\Firebase\Contract\Auth;
+use Kreait\Firebase\Contract\Transitional\FederatedUserFetcher;
 use Kreait\Firebase\Exception\Auth\InvalidOobCode;
 use Kreait\Firebase\Exception\Auth\RevokedIdToken;
 use Kreait\Firebase\Exception\Auth\RevokedSessionCookie;
@@ -518,6 +519,50 @@ abstract class AuthTestCase extends IntegrationTestCase
 
         $this->expectException(UserNotFound::class);
         $this->auth->getUserByPhoneNumber($phoneNumber);
+    }
+
+    #[Test]
+    public function getUserByProviderUid(): void
+    {
+        if (Util::authEmulatorHost() === null) {
+            $this->markTestSkipped('Getting user by provider UID can only be tested with the Firebase emulator.');
+        }
+
+        $auth = $this->auth;
+        if (!($auth instanceof FederatedUserFetcher)) {
+            $this->markTestSkipped('This test requires a FederatedUserFetcher implementation.');
+        }
+
+        $phoneNumber = '+1234567'.random_int(1000, 9999);
+
+        $user = $this->auth->createUser([
+            'phoneNumber' => $phoneNumber,
+        ]);
+
+        $check = $auth->getUserByProviderUid('phone', "$phoneNumber");
+
+        try {
+            $this->assertSame($user->uid, $check->uid);
+        } finally {
+            $auth->deleteUser($user->uid);
+        }
+
+    }
+
+    #[Test]
+    public function getUserByNonExistingProviderUid(): void
+    {
+        if (Util::authEmulatorHost() === null) {
+            $this->markTestSkipped('Getting user by provider UID can only be tested with the Firebase emulator.');
+        }
+
+        $auth = $this->auth;
+        if (!($auth instanceof FederatedUserFetcher)) {
+            $this->markTestSkipped('This test requires a FederatedUserFetcher implementation.');
+        }
+
+        $this->expectException(UserNotFound::class);
+        $auth->getUserByProviderUid('phone', '+192837465');
     }
 
     #[Test]